+++ R/elements.R
@@ -13,6 +13,7 @@
 #'     when it is disabled by a condition.
 #' @param height height of the glide (something like "400px" or "100%").
 #' @param keyboard set this to FALSE to disable keyboard navigation.
+#' @param swipe set this to FALSE to disable swipe navigation.
 #' @param custom_controls custom HTML or shiny tags to be used for the controls.
 #'     If `NULL``, use the default ones.
 #' @param controls_position either to place the default or custom controls on "top"
@@ -56,6 +57,7 @@ glide <- function(...,
   disable_type = c("disable", "hide"),
   height = "100%",
   keyboard = TRUE,
+  swipe = TRUE,
   custom_controls = NULL,
   controls_position = c("bottom", "top")) {
 
@@ -79,6 +81,7 @@ glide <- function(...,
   tagList(
     tags$div(class = "shinyglide", id = id,
             `data-keyboard` = keyboard,
+            `data-swipe` = swipe,
             `data-next-label` = next_label,
             `data-prev-label` = previous_label,
             `data-loading-label` = loading_label,