+++ srcjs/outputs/shinyglide.js
@@ -30,6 +30,7 @@ class ShinyGlide {
     this.loading_class = root.getAttribute("data-loading-class");
     this.disable_type = root.getAttribute("data-disable-type");
     this.keyboard = root.getAttribute("data-keyboard") != "FALSE";
+    this.disable_swipe = root.getAttribute("data-swipe") == "FALSE";
     this.prev_control = root.getElementsByClassName("prev-screen")[0];
     this.next_control = root.getElementsByClassName("next-screen")[0];
     this.first_control = root.getElementsByClassName("first-screen")[0];
@@ -77,6 +78,7 @@ class ShinyGlide {
     const glide = new Glide(this.root, {
       rewind: false,
       keyboard: this.keyboard,
+      swipeThreshold: this.disable_swipe ? false : 80,
       dragThreshold: false
     });